+++ src/Parser.js
@@ -25,10 +25,10 @@ import {
 
 import type {
   Attributes,
-  ParsedNodes,
+  ElementProps,
   NodeConfig,
   NodeInterface, // eslint-disable-line
-  ElementProps,
+  ReactNodeList,
 } from './types';
 
 const ELEMENT_NODE: number = 1;
@@ -39,7 +39,7 @@ const ARIA_COMPARE_LENGTH: number = 5;
 
 export default class Parser {
   doc: Document;
-  content: ParsedNodes;
+  content: ReactNodeList<*>;
   props: Object;
   matchers: Matcher<*>[];
   filters: Filter[];
@@ -84,7 +84,7 @@ export default class Parser {
   applyMatchers(
     string: string,
     parentConfig: NodeConfig,
-  ): string | Array<string | React.Element<*>> {
+  ): string | ReactNodeList<*> {
     const elements = [];
     const { props } = this;
     let matchedString = string;
@@ -373,7 +373,7 @@ export default class Parser {
    * while looping over all child nodes and generating an
    * array to interpolate into JSX.
    */
-  parse(): ParsedNodes {
+  parse(): ReactNodeList<*> {
     // $FlowIgnore Body is not null!
     return this.parseNode(this.doc.body, {
       ...CONFIG_BLOCK,
@@ -385,7 +385,7 @@ export default class Parser {
    * Loop over the nodes children and generate a
    * list of text nodes and React elements.
    */
-  parseNode(parentNode: NodeInterface, parentConfig: NodeConfig): ParsedNodes {
+  parseNode(parentNode: NodeInterface, parentConfig: NodeConfig): ReactNodeList<*> {
     const { noHtml, noHtmlExceptMatchers, disableWhitelist } = this.props;
     let content = [];
     let mergedText = '';